requirements.txt
Office365-REST-Python-Client
msal
requests
azure-identity
msgraph-sdk